--- a/saleor/order/models.py
+++ b/saleor/order/models.py
@@ -9,6 +9,7 @@ from django.core.validators import MinValueValidator
 from django.db import models
 from django.db.models import JSONField  # type: ignore
 from django.db.models import F, Max, Sum
+from django.db.models.expressions import Exists, OuterRef
 from django.utils.timezone import now
 from django_measurement.models import MeasurementField
 from django_prices.models import MoneyField, TaxedMoneyField
@@ -31,6 +32,7 @@ from ..payment.model_helpers import (
     get_total_authorized,
     get_total_captured,
 )
+from ..payment.models import Payment
 from ..shipping.models import ShippingMethod
 from . import FulfillmentStatus, OrderEvents, OrderStatus
 
@@ -59,9 +61,12 @@ class OrderQueryset(models.QuerySet):
         fulfilled).
         """
         statuses = {OrderStatus.UNFULFILLED, OrderStatus.PARTIALLY_FULFILLED}
-        qs = self.filter(status__in=statuses, payments__is_active=True)
+        payments = Payment.objects.filter(is_active=True).values("id")
+        qs = self.annotate(
+            matching=Exists(payments.filter(order_id=OuterRef("id")))
+        ).filter(matching=True)
         qs = qs.annotate(amount_paid=Sum("payments__captured_amount"))
-        return qs.filter(total_gross_amount__lte=F("amount_paid"))
+        return qs.filter(status__in=statuses, total_gross_amount__lte=F("amount_paid"))
 
     def ready_to_capture(self):
         """Return orders with payments to capture.
@@ -70,11 +75,14 @@ class OrderQueryset(models.QuerySet):
         have a preauthorized payment. The preauthorized payment can not
         already be partially or fully captured.
         """
-        qs = self.filter(
-            payments__is_active=True, payments__charge_status=ChargeStatus.NOT_CHARGED
-        )
-        qs = qs.exclude(status={OrderStatus.DRAFT, OrderStatus.CANCELED})
-        return qs.distinct()
+        payments = Payment.objects.filter(
+            is_active=True, charge_status=ChargeStatus.NOT_CHARGED
+        ).values("id")
+        qs = self.annotate(
+            matching=Exists(payments.filter(order_id=OuterRef("id")))
+        ).filter(matching=True)
+
+        return qs.exclude(status={OrderStatus.DRAFT, OrderStatus.CANCELED})
 
     def ready_to_confirm(self):
         """Return unconfirmed orders."""