I want to merge this change because it increases performance of order status filtering.
It implements subquery instead of doing distincts on large data sets.

Local testing [sample DB, 25k orders], querying orders with status on ["READY_TO_CAPTURE", "READY_TO_FULFILL"]:
Master: planning 4,98ms, execution: 118ms
This branch: planning 0,983ms, execution: 10,4ms

<!-- Please mention all relevant issue numbers. -->

# Impact

* [ ] New migrations
* [ ] New/Updated API fields or mutations
* [ ] Deprecated API fields or mutations
* [ ] Removed API types, fields, or mutations
* [ ] Documentation needs to be updated

# Pull Request Checklist

<!-- Please keep this section. It will make the maintainer's life easier. -->

* [ ] Privileged queries and mutations are guarded by proper permission checks
* [ ] Database queries are optimized and the number of queries is constant
* [ ] Database migration files are up to date
* [X] The changes are tested
* [ ] GraphQL schema and type definitions are up to date
* [ ] Changes are mentioned in the changelog