@patrys 2021-04-28
Task linked: [SALEOR-3106 Optimize Order filtering by OrderStatus=READY_TO_CAPTURE](https://app.clickup.com/t/2549495/SALEOR-3106)